phoenix/requirements.txt
arize-phoenix-otel>=0.6.1
openai>=1.0.0
python-dotenv>=1.0.0
tiktoken>=0.6.0
chromadb>=0.4.0
langchain>=0.1.0
langchain-community>=0.0.38
langchain-core>=0.1.53
sentence-transformers>=2.2.2
openinference-semantic-conventions>=0.1.12